feat: support passing in extractors themselves
