+++ cmd/osv-scanner/scan/source/command.go
@@ -129,7 +129,7 @@ func action(_ context.Context, cmd *cli.Command, stdout, stderr io.Writer) error
 		ExperimentalScannerActions: experimentalScannerActions,
 	}
 
-	if len(experimentalScannerActions.ExtractorNames) == 0 {
+	if len(experimentalScannerActions.Extractors) == 0 {
 		return errors.New("at least one extractor must be enabled")
 	}
